+++ framy_generator/test/generators/widget_cutom_page_template_generator_test.dart
@@ -4,28 +4,28 @@ import 'package:test/test.dart';
 void main() {
   group('generateCustomPage result', () {
     test('should contain LayoutBuilder', () {
-      final result = generateCustomPage();
+      final result = generateCustomPage(false);
       expect(result.contains('LayoutBuilder('), isTrue);
     });
 
     test('should contain list with dependencies', () {
-      final result = generateCustomPage();
+      final result = generateCustomPage(false);
       expect(
           result.contains('List<FramyDependencyModel> dependencies'), isTrue);
     });
 
     test('should contain FramyWidgetDependenciesPanel', () {
-      final result = generateCustomPage();
+      final result = generateCustomPage(false);
       expect(result.contains('FramyWidgetDependenciesPanel('), isTrue);
     });
 
     test('should contain FramyWidgetDependenciesFAB', () {
-      final result = generateCustomPage();
+      final result = generateCustomPage(false);
       expect(result.contains('FramyWidgetDependenciesFAB('), isTrue);
     });
 
     test('should have an initialized presets field', () {
-      final result = generateCustomPage();
+      final result = generateCustomPage(false);
       expect(
         result.contains(
             'final Map<String, Map<String, dynamic>> presets = createFramyPresets();'),
@@ -34,10 +34,26 @@ void main() {
     });
 
     test('should contain definition of onChanged method', () {
-      final result = generateCustomPage();
+      final result = generateCustomPage(false);
       expect(
         result.contains('void onChanged(FramyDependencyModel dependencyModel)'),
         isTrue,
+      );
+    });
+
+    test('should contain DevicePreview when true as param is passed', () {
+      final result = generateCustomPage(true);
+      expect(
+        result.contains('DevicePreview('),
+        isTrue,
+      );
+    });
+
+    test('should not contain DevicePreview when false as param is passed', () {
+      final result = generateCustomPage(false);
+      expect(
+        result.contains('DevicePreview('),
+        isFalse,
       );
     });
   });