+++ test_apps/weight_tracker/lib/main.dart
@@ -7,7 +7,7 @@ void main() {
   runApp(WeightTrackerApp());
 }
 
-@framyApp
+@FramyApp(useDevicePreview: true)
 class WeightTrackerApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {